+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - React Redux and Redux toolkit
 - Ability to select an image with Expo Image Picker
 - A component to show the selected image
+- The Create Tab for making sliced images (slicing not implemented)
+- A FlatList image viewer that snap to 1/3 of the image width
 
 ### Changed